Set indent to four spaces for Markdown
This is how much code blocks need to be indented in Markdown.

It's also how much you need to indent in order to continue a list item
after a blank line, in the original Markdown and in Python-Markdown.

And 8 spaces is how much you need to indent to put a code block in a list
item.

So four spaces definitely seems right here. There's never any reason to
indent anything by just two spaces in Markdown.